+++ src/commands/check/index.ts
@@ -4,7 +4,6 @@ import { detectIacFormat } from './detect-iac-format';
 import { getConfig } from './get-config';
 import { prepareForCheck } from './prepare';
 import { checkTemplates, testResource } from './checks';
-import { writeFileSync, existsSync } from 'fs';
 
 async function check (options: CheckOptions) {
   const config = getConfig(options);
@@ -30,38 +29,6 @@ async function check (options: CheckOptions) {
   errors.forEach(logger.cliError, logger);
 }
 
-async function init () { 
-  const configData = `
-  {
-    "awsCdkParsers": [
-        "@tinystacks/aws-cdk-parser"
-    ],
-    "terraformParsers": [
-        "@tinystacks/terraform-resource-parser",
-        "@tinystacks/terraform-module-parser"
-    ],
-    "templateChecks": [
-        "@tinystacks/aws-template-checks"  
-    ],
-    "resourceChecks": [
-        "@tinystacks/aws-resource-checks"
-    ]
-  }
-  `;
-  const dirname = './precloud.config.json';
-  if (existsSync(dirname)){
-    logger.info('Configuration file already exists, not creating a default one');
-    return;
-  }
-  try { 
-    writeFileSync(dirname, configData);
-    logger.success('Configuration file successfully created!');
-  } catch(e) { 
-    logger.error(`Error creating configuration file: ${e}`);
-  }
-}
-
 export {
-  check,
-  init
+  check
 };
