+++ src/commands/index.ts
@@ -1,1 +1,2 @@
 export * from './check';
+export * from './init';
