+++ test/commands/check/checks/aws/MockResourceChecks.ts
