+++ test/commands/check/checks/aws/MockTemplateChecks.ts
