+++ test/commands/check/checks/aws/index.test.ts
